Removed unnecessary type cast for current_ping in cCockpit.cpp ('float => int => string' instead of 'float => int => float => string')

+++ src/tron/cockpit/cCockpit.cpp
@@ -198,7 +198,7 @@ tValue::BasePtr cCockpit::cb_CurrentAcceleration(void) {
 }
 tValue::BasePtr cCockpit::cb_CurrentPing(void) {
     if(m_Type == VIEWPORT_TOP) return tValue::BasePtr(new tValue::Base());
-    return tValue::BasePtr(new tValue::Float(static_cast<int>(m_ViewportPlayer->ping*1000)));
+    return tValue::BasePtr(new tValue::Int(m_ViewportPlayer->ping*1000));
 }
 tValue::BasePtr cCockpit::cb_CurrentSpeed(void) {
     if(m_Type == VIEWPORT_TOP) return tValue::BasePtr(new tValue::Base());
